pyproject.toml
[tool.poetry]
name = "y"
version = "0.1.0"
description = "Utilities to count cells faster"
authors = ["Michael Mezzina <michaelmez39@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.10"
tensorflow = "~2.8.0"
numpy = "^1.22.2"
pillow = "~9.0.1"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
